+++ R/grid_code_paths.R
@@ -225,6 +225,11 @@ tune_grid_loop_impl <- function(fn_tune_grid_loop_iter,
 
     seeds <- generate_seeds(rng, n_splits * n_grid_info)
 
+    # Make `slice_seeds` available in the environment that foreach will
+    # search for it in; resolves issue with PSOCK/multisession parallelism
+    # (#888, #889).
+    slice_seeds <- slice_seeds
+
     # Evaluate the call to foreach in a local environment using a clone of
     # the current environment so that foreach doesn't touch the exit
     # handlers attached to the execution environment of this function
@@ -243,7 +248,6 @@ tune_grid_loop_impl <- function(fn_tune_grid_loop_iter,
           foreach::foreach(
             row = rows,
             .combine = iter_combine,
-            seed = slice_seeds(seeds, iteration, n_grid_info),
             .options.future = list(seed = NULL, packages = packages)
           )
       } else {
@@ -257,7 +261,6 @@ tune_grid_loop_impl <- function(fn_tune_grid_loop_iter,
           foreach::foreach(
             row = rows,
             .combine = iter_combine,
-            seed = slice_seeds(seeds, iteration, n_grid_info),
             .packages = packages,
             .errorhandling = "pass"
           )
@@ -266,6 +269,7 @@ tune_grid_loop_impl <- function(fn_tune_grid_loop_iter,
       suppressPackageStartupMessages(
         for_each %op% {
             grid_info_row <- vctrs::vec_slice(grid_info, row)
+            seed <- slice_seeds(seeds, iteration, n_grid_info)[[1]]
 
             # Likely want to debug with `debugonce(tune_grid_loop_iter)`
             fn_tune_grid_loop_iter_safely(