+++ tests/testthat/helper-tune-package.R
@@ -26,3 +26,46 @@ helper_objects_tune <- function() {
     svm_mod = svm_mod
   )
 }
+
+# When not interactive, the cli progress bar is occasionally "reset" at random
+# intervals throughout tuning, which means that intermediate counts may appear
+# in snapshots. In an output like:
+#
+# > A | error:   AHHhH
+# There were issues with some computations   A: x2
+# There were issues with some computations   A: x4
+# There were issues with some computations   A: x5
+# There were issues with some computations   A: x5
+#
+# ...we just want to see the unique issues, e.g.
+#
+# > A | error:   AHHhH
+catalog_lines <- function(lines) {
+  lines[grepl("^>", lines)]
+}
+
+# Make a new binding to prevent infinite recursion when the original is mocked.
+initialize_catalog_ <- tune:::initialize_catalog
+
+# Sets a new exit handler on `initialize_catalog()` that stores the summary
+# of issues before it's cleared along with the progress bar. Together with
+# the above, we can test the full catalog output.
+redefer_initialize_catalog <- function(test_env) {
+  local({
+    function(control, env = rlang::caller_env()) {
+      initialize_catalog_(control, env)
+
+      withr::defer(
+        assign(
+          "catalog_summary_test",
+          tune:::tune_env$progress_env$catalog_summary,
+          test_env
+        ),
+        envir = env,
+        priority = "first"
+      )
+
+      NULL
+    }
+  })
+}