+++ tests/testthat/test-logging.R
@@ -38,21 +38,33 @@ test_that("log issues", {
   note_2 <- tibble::tibble(location = "toledo", type = "error", note = "Error in log(\"a\"): non-numeric argument to mathematical function")
 
   expect_snapshot(
-    expect_equal(
-      tune:::log_problems(note_1, ctrl_f, rs, "toledo", res_1, bad_only = FALSE),
-      dplyr::bind_rows(note_1, note_2)
-    )
+    problems_1 <-
+      tune:::log_problems(note_1, ctrl_f, rs, "toledo", res_1, bad_only = FALSE)
   )
+
+  expect_equal(
+    dplyr::select(problems_1, -trace),
+    dplyr::bind_rows(note_1, note_2)
+  )
+
+  expect_null(problems_1$trace[[1]])
+  expect_s3_class(problems_1$trace[[2]], "rlang_trace")
 
   expect_silent(tune:::log_problems(note_1, ctrl_f, rs, "toledo", res_2, bad_only = FALSE))
 
   note_3 <- tibble::tibble(location = "toledo", type = "warning", note = "NaNs produced")
   expect_snapshot(
-    expect_equal(
-      tune:::log_problems(note_1, ctrl_f, rs, "toledo", res_3, bad_only = FALSE),
-      dplyr::bind_rows(note_1, note_3)
-    )
+    problems_2 <-
+      tune:::log_problems(note_1, ctrl_f, rs, "toledo", res_3, bad_only = FALSE)
   )
+
+  expect_equal(
+    dplyr::select(problems_2, -trace),
+    dplyr::bind_rows(note_1, note_3)
+  )
+
+  expect_null(problems_2$trace[[1]])
+  expect_s3_class(problems_2$trace[[2]], "rlang_trace")
 })
 
 
@@ -198,4 +210,271 @@ test_that("message_wrap", {
     ),
     crayon = TRUE
   )
+})
+
+test_that("interactive logger works (fit_resamples, warning + error)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+
+  raise_warning <- function(x) {warning("ope! yikes.")}
+  raise_error <- function(x) {stop("AHHhH")}
+
+  set.seed(1)
+  expect_snapshot(
+    {res_fit <-
+      fit_resamples(
+        parsnip::nearest_neighbor("regression", "kknn"),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 5),
+        control = control_resamples(
+          extract = function(x) {raise_warning(); raise_error()})
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
+})
+
+test_that("interactive logger works (fit_resamples, rlang warning + error)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+
+  raise_warning_rl <- function(x) {rlang::warn("ope! yikes. (but rlang)")}
+  raise_error_rl <- function(x) {rlang::abort("AHHhH (but rlang)")}
+
+  set.seed(1)
+  expect_snapshot(
+    {res_fit <-
+      fit_resamples(
+        parsnip::nearest_neighbor("regression", "kknn"),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 5),
+        control = control_resamples(
+          extract = function(x) {raise_warning_rl(); raise_error_rl()}
+        )
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
+})
+
+
+test_that("interactive logger works (fit_resamples, multiline)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+  skip_on_cran()
+
+  raise_multiline_conditions <- function(x) {
+    cli::cli_warn(c("hmmm what's happening", "uuuhhHhH"))
+    cli::cli_abort(c("aHHHksdjvndiuf", "!" = "),:"))
+    x
+  }
+
+  set.seed(1)
+  expect_snapshot(
+    {res_fit <-
+      fit_resamples(
+        parsnip::nearest_neighbor("regression", "kknn"),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 5),
+        control = control_resamples(extract = raise_multiline_conditions)
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
+})
+
+test_that("interactive logger works (fit_resamples, occasional error)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+  skip_on_cran()
+
+  raise_error_later <- function() {local({
+    count <- 0
+    function(x) {
+      count <<- count + 1
+      if (count > 3) {
+        stop("this errors now! ha!")
+      }
+      "hi"
+    }
+  })}
+  later <- raise_error_later()
+
+  set.seed(1)
+  expect_snapshot(
+    {res_fit <-
+      fit_resamples(
+        parsnip::nearest_neighbor("regression", "kknn"),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 5),
+        control = control_resamples(extract = later)
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
+})
+
+test_that("interactive logger works (fit_resamples, occasional errors)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+
+  skip_on_cran()
+
+  raise_error_once <- function() {local({
+    first <- TRUE
+    function(x) {
+      if (first) {
+        first <<- FALSE
+        stop("oh no")
+      }
+
+      "hi"
+    }
+  })}
+
+  raise_error_later <- function() {local({
+    count <- 0
+    function(x) {
+      count <<- count + 1
+      if (count > 3) {
+        stop("this errors now! ha!")
+      }
+      "hi"
+    }
+  })}
+
+  once <- raise_error_once()
+  later <- raise_error_later()
+
+  set.seed(1)
+  expect_snapshot(
+    {res_fit <-
+      fit_resamples(
+        parsnip::nearest_neighbor("regression", "kknn"),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 10),
+        control = control_resamples(extract = function(x) {once(); later()})
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
+})
+
+
+test_that("interactive logger works (fit_resamples, many distinct errors)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+
+  skip_on_cran()
+  # Enough characters to see 'E: x1'
+  rlang::local_options(cli.width = 84)
+
+  raise_error_numbered <- function() {local({
+    count <- 0
+    function(x) {
+      count <<- count + 1
+      stop(paste0("error number ", count))
+      "hi"
+    }
+  })}
+  numbered <- raise_error_numbered()
+
+  set.seed(1)
+  expect_snapshot(
+    {res_fit <-
+      fit_resamples(
+        parsnip::nearest_neighbor("regression", "kknn"),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 5),
+        control = control_resamples(extract = numbered)
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
+})
+
+test_that("interactive logger works (tune grid, error)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+  skip_on_cran()
+
+  raise_error <- function(x) {stop("AHHhH")}
+
+  set.seed(1)
+  expect_snapshot(
+    {res_fit <-
+      tune_grid(
+        parsnip::nearest_neighbor("regression", "kknn", dist_power = tune()),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 5),
+        grid = 5,
+        control = control_grid(extract = raise_error)
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
+})
+
+test_that("interactive logger works (bayesian, error)", {
+  skip_if(tune:::allow_parallelism(FALSE), "Will not catalog: parallelism is enabled")
+  local_mocked_bindings(
+    is_testing = function() {FALSE},
+    initialize_catalog = redefer_initialize_catalog(rlang::current_env())
+  )
+
+  skip_on_cran()
+
+  raise_error <- function(x) {stop("AHHhH")}
+
+  set.seed(1)
+  expect_snapshot(
+    {res_grid <-
+      tune_bayes(
+        parsnip::nearest_neighbor("regression", "kknn", dist_power = tune()),
+        Sale_Price ~ .,
+        rsample::vfold_cv(modeldata::ames[, c(72, 40:45)], 5),
+        initial = 5,
+        iter = 5,
+        control = control_bayes(extract = raise_error)
+    )},
+    transform = catalog_lines
+  )
+
+  # `catalog_summary_test` written to this env via `redefer_initialize_catalog()`
+  expect_snapshot(catalog_summary_test)
 })